# api/openapi.yaml
openapi: 3.1.0
info:
  title: tyrs-hand Management API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
...
paths:
...
  /trigger-rules:
    get: { $ref: "#/components/pathItems/ListTriggerRules" }
    post:
      operationId: createTriggerRule
      tags: [Resources]
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TriggerRuleInput" }
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IDResource" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tyrs_hand_session
    nodeBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-node-token
    clientBearer:
      type: http
      scheme: bearer
      bearerFormat: tyrs-hand-device-token
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    ResourceList:
      description: 游标分页资源列表
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ResourceList" }
  pathItems:
...
    ListTriggerRules:
      operationId: listTriggerRules
      tags: [Resources]
      parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { $ref: "#/components/responses/ResourceList" }, default: { $ref: "#/components/responses/Problem" } }
...
  schemas:
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        requestId: { type: string }
...
    TriggerRuleInput:
      type: object
      required: [repositoryId, agentProfileId, name, eventName, triggerKind, instruction]
      properties:
        repositoryId: { type: string, format: uuid }
        agentProfileId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 128 }
        eventName: { type: string, minLength: 1 }
        action: { type: string }
        enabled: { type: boolean }
        priority: { type: integer, default: 100 }
        triggerKind:
          type: string
          enum: [event, label, slash_command, mention_command, legacy_mention]
        triggerValue:
          type: string
          description: Label 名或不带斜杠的命令名；event、mention_command 与 legacy_mention 留空。
        actorMinPermission:
          type: string
          enum: [read, triage, write, maintain, admin]
          default: triage
        instruction: { type: string, minLength: 1 }
        skills: { type: array, items: { type: string } }
        allowedTools: { type: array, items: { type: string } }
        dangerousActions: { type: array, items: { type: string } }
        filters: { type: object, additionalProperties: true }
...
    IDResource:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
    ResourceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { type: object, additionalProperties: true }
        nextCursor: { type: string }
security:
  - sessionCookie: []